+++ src/cryptonote_core/tx_pool.cpp
@@ -1332,53 +1332,6 @@ namespace cryptonote
     return true;
   }
   //---------------------------------------------------------------------------------
-  bool tx_memory_pool::get_pool_for_rpc(std::vector<cryptonote::rpc::tx_in_pool>& tx_infos, cryptonote::rpc::key_images_with_tx_hashes& key_image_infos) const
-  {
-    auto locks = tools::unique_locks(m_transactions_lock, m_blockchain);
-
-    tx_infos.reserve(m_blockchain.get_txpool_tx_count());
-    key_image_infos.reserve(m_blockchain.get_txpool_tx_count());
-    m_blockchain.for_all_txpool_txes([&tx_infos, key_image_infos](const crypto::hash &txid, const txpool_tx_meta_t &meta, const cryptonote::blobdata *bd){
-      cryptonote::rpc::tx_in_pool txi;
-      txi.tx_hash = txid;
-      if (!parse_and_validate_tx_from_blob(*bd, txi.tx))
-      {
-        MERROR("Failed to parse tx from txpool");
-        // continue
-        return true;
-      }
-      txi.tx.set_hash(txid);
-      txi.blob_size = bd->size();
-      txi.weight = meta.weight;
-      txi.fee = meta.fee;
-      txi.kept_by_block = meta.kept_by_block;
-      txi.max_used_block_height = meta.max_used_block_height;
-      txi.max_used_block_hash = meta.max_used_block_id;
-      txi.last_failed_block_height = meta.last_failed_height;
-      txi.last_failed_block_hash = meta.last_failed_id;
-      txi.receive_time = meta.receive_time;
-      txi.relayed = meta.relayed;
-      txi.last_relayed_time = meta.last_relayed_time;
-      txi.do_not_relay = meta.do_not_relay;
-      txi.double_spend_seen = meta.double_spend_seen;
-      tx_infos.push_back(txi);
-      return true;
-    }, true, false);
-
-    for (const key_images_container::value_type& kee : m_spent_key_images) {
-      std::vector<crypto::hash> tx_hashes;
-      const std::unordered_set<crypto::hash>& kei_image_set = kee.second;
-      for (const crypto::hash& tx_id_hash : kei_image_set)
-      {
-        tx_hashes.push_back(tx_id_hash);
-      }
-
-      const crypto::key_image& k_image = kee.first;
-      key_image_infos[k_image] = std::move(tx_hashes);
-    }
-    return true;
-  }
-  //---------------------------------------------------------------------------------
   bool tx_memory_pool::check_for_key_images(const std::vector<crypto::key_image>& key_images, std::vector<bool>& spent) const
   {
     auto locks = tools::unique_locks(m_transactions_lock, m_blockchain);
